Fix csv extractor
